Enlarge attributes to 32 bit instead of 16. This prepares 256-color support for terminal.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@38089 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/terminal/BasicTerminalBuffer.h
@@ -77,10 +77,10 @@ public:
 			bool				IsFullWidthChar(int32 row, int32 column) const;
 			int					GetChar(int32 row, int32 column,
 									UTF8Char& character,
-									uint16& attributes) const;
+									uint32& attributes) const;
 			int32				GetString(int32 row, int32 firstColumn,
 									int32 lastColumn, char* buffer,
-									uint16& attributes) const;
+									uint32& attributes) const;
 			void				GetStringFromRegion(BString& string,
 									const TermPos& start,
 									const TermPos& end) const;
