+++ src/apps/terminal/TermView.h
@@ -139,7 +139,7 @@ private:
 			void			_Activate();
 			void			_Deactivate();
 
-			void			_DrawLinePart(int32 x1, int32 y1, uint16 attr,
+			void			_DrawLinePart(int32 x1, int32 y1, uint32 attr,
 								char* buffer, int32 width, bool mouse,
 								bool cursor, BView* inView);
 			void			_DrawCursor();
